Fix roadmap sequencing and decision-sheet drift from architecture review
Replaces the roadmap with the 8-phase (0-7) structure that groups
Backtesting+Experiment Tracking and Risk Engine+Paper Trading together,
and updates CLAUDE.md/AGENTS.md to mirror it as the authoritative
implementation order (previously the roadmap silently omitted a
distinct Risk & Metrics phase that both governance docs called out).

Also brings the already-committed ingestion code back in line with the
decisions in docs/answers_obsidian_optimized.md, which the pre-implementation
review found it had drifted from:

- candles/funding_rates price, volume, and rate columns now NUMERIC(28,12)
  instead of DOUBLE PRECISION (Q10), with CandleRecord/FundingRecord using
  Decimal end-to-end so precision isn't lost before the DB layer
- BybitExchange.get_candles now takes the canonical timeframe vocabulary
  (1m/5m/15m/1h/4h/1d, Q11) as its public parameter and translates to
  Bybit's native interval codes internally, instead of leaking Bybit's
  own codes into the stored `timeframe` column (which broke cross-exchange
  joins on the same nominal candle)

+++ backend/data/adapters/binance.py
@@ -9,6 +9,7 @@ Docs: https://binance-docs.github.io/apidocs/spot/en/#kline-candlestick-data
 """
 
 from datetime import datetime, timezone
+from decimal import Decimal
 
 import httpx
 
@@ -62,11 +63,11 @@ class BinanceExchange(Exchange):
                     timeframe=timeframe,
                     open_time=_ms_to_utc(open_ms),
                     close_time=_ms_to_utc(close_ms),
-                    open=float(open_),
-                    high=float(high),
-                    low=float(low),
-                    close=float(close),
-                    volume=float(volume),
+                    open=Decimal(open_),
+                    high=Decimal(high),
+                    low=Decimal(low),
+                    close=Decimal(close),
+                    volume=Decimal(volume),
                     ingestion_time=now,
                     raw_payload={"row": row},
                 )
